Update Game.Catch_the_ball - ex. 7
Before starting the game, create a file with a name 'Leaderboard.txt' in the same folder with the game.
To start the game you need to enter a nickname. To view the leaderboard you have to click the leaderboard button and hold it. 
The game goes on for 20 seconds. By clicking on a ball you get 1 point, but you get 10 points from each star. Stars change their velocity randomly every interval of time.